test(parser): split grammar/tests/mod.rs into themed child modules
Issue #391 item 1: extract the oversized src/parser/grammar/tests/mod.rs
(2848 -> 71 lines) into 8 new themed child test modules, continuing the
existing grammar/tests themed-file pattern. Pure test code motion: all 156
parser::grammar::tests functions unchanged (same count before and after);
mod.rs retains only shared helpers and mod declarations.

No behavior change: full suite green (1608), clippy/fmt clean both feature sets.
